Test Harness: integrated initial support for report compiler.
- Test Harness now outputs the file `benchmark_list.txt` with the list
of report filenames generated during the run, compatible with report
compiler. Filenames are relative to the list location. List file is
writen to the output directory.

- Modified Test Harness to be able to compute the simplified, pretty-
print run overview locally since the functinality was moved to report
compiler for heavier statistical computations.

- Removed capability to select report timing units from test harness
since this is now part of the report compiler.