Separated Report Compiler into an executable and a shared library.
- Report compiler core functionality has been moved to a shared library to
allow for external programs to use it without having to call an executable.

- Report compiler executable now simply parses the command line and calls
the shared library core functionality for operation. The only way to do
report compilation using custom user options is to use this executable
(or to link into the shared library and call the core functionality with
custom configuration).

- Test Harness performs report compilation by default by calling into the
report compiler's shared library using default compiler options.