Dataset loader (CSV): revamped to support string inputs.
- Revamped CSV dataset loader to use a CSV tokenizer instead of C++
streams. This should add more flexibility to parsing as each CSV
line can be directly decomposed into string tokens.

- Added support for users to use strings as values for dataset samples.
If loader is unable to load a value in the specified type, it will
attempt to load it as a collection of bytes and cast each byte to
the data type.

- Users can add padding to their samples by leaving empty elements in
a CSV line. The loader will interpret these empty elements as "fill out
the rest of the elements with padding." This should help with avoiding
users needing to repeat padding manually in their datasets. Users can
take advantage of this feature to add extra dimensions to their input
data by maing inner dimensions of the same size and padding to fill
each entry in the sample.

- Removed checking for correct number of inputs and outputs, as well as
number of elements in a sample. The logic was to have the loader be
more flexible to allow loading without many restriction. The validation
is left to the client of the loader, to check that the correct sizes,
inputs and outputs were specified in the dataset.